+++ src/main/java/edu/wpi/N/entities/memento/GlobalKeyListener.java
@@ -1,12 +1,7 @@
 package edu.wpi.N.entities.memento;
 
 import edu.wpi.N.App;
-import edu.wpi.N.database.DBException;
 import edu.wpi.N.entities.States.StateSingleton;
-import java.io.IOException;
-import java.util.Timer;
-import java.util.TimerTask;
-import javafx.application.Platform;
 import org.jnativehook.GlobalScreen;
 import org.jnativehook.NativeHookException;
 import org.jnativehook.keyboard.NativeKeyEvent;
@@ -26,7 +21,7 @@ public class GlobalKeyListener implements NativeKeyListener {
   }
 
   public void nativeKeyPressed(NativeKeyEvent e) {
-    System.out.println("Key Pressed: " + NativeKeyEvent.getKeyText(e.getKeyCode()));
+    // System.out.println("Key Pressed: " + NativeKeyEvent.getKeyText(e.getKeyCode()));
 
     if (e.getKeyCode() == NativeKeyEvent.VC_ESCAPE) {
       try {
@@ -37,7 +32,8 @@ public class GlobalKeyListener implements NativeKeyListener {
     }
 
     try {
-      update();
+      StateSingleton singleton = StateSingleton.getInstance();
+      singleton.update();
     } catch (Exception ex) {
       ex.printStackTrace();
       System.out.println("Failed at GlobalKeyListener - nativeKeyPressed()");
@@ -47,7 +43,8 @@ public class GlobalKeyListener implements NativeKeyListener {
   public void nativeKeyReleased(NativeKeyEvent e) {
     // System.out.println("Key Released: " + NativeKeyEvent.getKeyText(e.getKeyCode()));
     try {
-      update();
+      StateSingleton singleton = StateSingleton.getInstance();
+      singleton.update();
     } catch (Exception ex) {
       ex.printStackTrace();
       System.out.println("Failed at GlobalKeyListener - nativeKeyReleased()");
@@ -57,45 +54,18 @@ public class GlobalKeyListener implements NativeKeyListener {
   public void nativeKeyTyped(NativeKeyEvent e) {
     // System.out.println("Key Typed: " + e.getKeyText(e.getKeyCode()));
     try {
-      update();
+      StateSingleton singleton = StateSingleton.getInstance();
+      singleton.update();
     } catch (Exception ex) {
       ex.printStackTrace();
       System.out.println("Failed at GlobalKeyListener - nativeKeyTyped()");
     }
   }
 
-  public void update() throws DBException {
-    StateSingleton singleton = StateSingleton.getInstance();
-    singleton.timer.purge();
-    singleton.timer.cancel();
-    singleton.timer = new Timer();
-    TimerTask timerTask =
-        new TimerTask() {
-          @Override
-          public void run() {
-            Platform.runLater(
-                () -> {
-                  System.out.println("Timer Ended!");
-                  System.out.println("Reset Kiosk!");
-                  try {
-                    singleton.originator.getStateFromMemento(singleton.careTaker.get(0));
-                    String path = singleton.originator.getState();
-                    switchTheScene(path);
-                  } catch (Exception e) {
-                    e.printStackTrace();
-                    System.out.println("Why u no work? 1");
-                  }
-                });
-          }
-        };
-
-    singleton.timer.schedule(timerTask, singleton.timeoutTime);
-  }
-
-  public void switchTheScene(String path) throws IOException, DBException {
-    StateSingleton singleton = StateSingleton.getInstance();
-    singleton.timer.cancel();
-    singleton.timer.purge();
-    mainApp.switchScene(path, singleton);
-  }
+  //  public void switchTheScene(String path) throws IOException, DBException {
+  //    StateSingleton singleton = StateSingleton.getInstance();
+  //    singleton.timer.cancel();
+  //    singleton.timer.purge();
+  //    mainApp.switchScene(path, singleton);
+  //  }
 }
