+++ src/main/java/edu/wpi/N/entities/memento/GlobalMouseListener.java
@@ -1,12 +1,7 @@
 package edu.wpi.N.entities.memento;
 
 import edu.wpi.N.App;
-import edu.wpi.N.database.DBException;
 import edu.wpi.N.entities.States.StateSingleton;
-import java.io.IOException;
-import java.util.Timer;
-import java.util.TimerTask;
-import javafx.application.Platform;
 import org.jnativehook.mouse.NativeMouseEvent;
 import org.jnativehook.mouse.NativeMouseInputListener;
 
@@ -24,9 +19,10 @@ public class GlobalMouseListener implements NativeMouseInputListener {
   }
 
   public void nativeMouseClicked(NativeMouseEvent e) {
-    // System.out.println("Mouse Clicked: " + e.getClickCount());
+    System.out.println("Mouse Clicked: " + e.getClickCount());
     try {
-      update();
+      StateSingleton singleton = StateSingleton.getInstance();
+      singleton.update();
     } catch (Exception ex) {
       ex.printStackTrace();
       System.out.println("Failed at GlobalMouseListener - nativeMouseClicked()");
@@ -34,9 +30,10 @@ public class GlobalMouseListener implements NativeMouseInputListener {
   }
 
   public void nativeMousePressed(NativeMouseEvent e) {
-    // System.out.println("Mouse Pressed: " + e.getButton());
+    System.out.println("Mouse Pressed: " + e.getButton());
     try {
-      update();
+      StateSingleton singleton = StateSingleton.getInstance();
+      singleton.update();
     } catch (Exception ex) {
       ex.printStackTrace();
       System.out.println("Failed at GlobalMouseListener - nativeMousePressed()");
@@ -44,9 +41,10 @@ public class GlobalMouseListener implements NativeMouseInputListener {
   }
 
   public void nativeMouseReleased(NativeMouseEvent e) {
-    // System.out.println("Mouse Released: " + e.getButton());
+    System.out.println("Mouse Released: " + e.getButton());
     try {
-      update();
+      StateSingleton singleton = StateSingleton.getInstance();
+      singleton.update();
     } catch (Exception ex) {
       ex.printStackTrace();
       System.out.println("Failed at GlobalMouseListener - nativeMouseReleased()");
@@ -54,9 +52,10 @@ public class GlobalMouseListener implements NativeMouseInputListener {
   }
 
   public void nativeMouseMoved(NativeMouseEvent e) {
-    // System.out.println("Mouse Moved: " + e.getX() + ", " + e.getY());
+    System.out.println("Mouse Moved: " + e.getX() + ", " + e.getY());
     try {
-      update();
+      StateSingleton singleton = StateSingleton.getInstance();
+      singleton.update();
     } catch (Exception ex) {
       ex.printStackTrace();
       System.out.println("Failed at GlobalMouseListener - nativeMouseMoved()");
@@ -64,47 +63,20 @@ public class GlobalMouseListener implements NativeMouseInputListener {
   }
 
   public void nativeMouseDragged(NativeMouseEvent e) {
-    // System.out.println("Mouse Dragged: " + e.getX() + ", " + e.getY());
+    System.out.println("Mouse Dragged: " + e.getX() + ", " + e.getY());
     try {
-      update();
+      StateSingleton singleton = StateSingleton.getInstance();
+      singleton.update();
     } catch (Exception ex) {
       ex.printStackTrace();
       System.out.println("Failed at GlobalMouseListener - nativeMouseDragged()");
     }
   }
 
-  public void update() throws DBException {
-    StateSingleton singleton = StateSingleton.getInstance();
-    singleton.timer.purge();
-    singleton.timer.cancel();
-    singleton.timer = new Timer();
-    TimerTask timerTask =
-        new TimerTask() {
-          @Override
-          public void run() {
-            Platform.runLater(
-                () -> {
-                  System.out.println("Timer Ended!");
-                  System.out.println("Reset Kiosk!");
-                  try {
-                    singleton.originator.getStateFromMemento(singleton.careTaker.get(0));
-                    String path = singleton.originator.getState();
-                    switchTheScene(path);
-                  } catch (Exception e) {
-                    e.printStackTrace();
-                    System.out.println("Why u no work? 2");
-                  }
-                });
-          }
-        };
-
-    singleton.timer.schedule(timerTask, singleton.timeoutTime);
-  }
-
-  public void switchTheScene(String path) throws IOException, DBException {
-    StateSingleton singleton = StateSingleton.getInstance();
-    singleton.timer.cancel();
-    singleton.timer.purge();
-    mainApp.switchScene(path, singleton);
-  }
+  //  public void switchTheScene(String path) throws IOException, DBException {
+  //    StateSingleton singleton = StateSingleton.getInstance();
+  //    singleton.timer.cancel();
+  //    singleton.timer.purge();
+  //    mainApp.switchScene(path, singleton);
+  //  }
 }
